requirements.txt
azure-identity
python-dotenv
langchain-core
langchain-community
langchain-openai
langchain
streamlit